Refactored jc69 to separate module
